Restore separate menuitems for opening tabs and windows
Without the separate menuitems, it was not possible to open new tabs or
windows solely by using the mouse if the preference was set to windows
or tabs respectively. This isn't ideal for accessibility.

The preference is also hard to discover and there was simply no strong
justification for unifying the menuitems in the first place.

The org.gnome.Terminal.Legacy.Settings:new-terminal-mode preference is
now ignored.

This effectively reverts 99fc0136a5be6323b81b8b339482bc699b53e1f9 and
1fecaa3685a8a4d4e7027fc6c9e60df2ca69f5b6

https://bugzilla.redhat.com/show_bug.cgi?id=1300826