Restore separate menuitems for opening tabs and windows
Without the separate menuitems, it was not possible to open new tabs or
windows solely by using the mouse if the preference was set to windows
or tabs respectively. This isn't ideal for accessibility.

The preference is also hard to discover and there was simply no strong
justification for unifying the menuitems in the first place.

The org.gnome.Terminal.Legacy.Settings:new-terminal-mode preference is
now ignored.

https://bugzilla.redhat.com/show_bug.cgi?id=1300826