screen: Track the current foreground process
... using the interactive shell's precmd and preexec hooks, by
monitoring the contents of the terminal, and a combination of
tcgetpgrp(3) and /proc, instead of relying on the shell's history.

The shell's history cannot be reliably used to find out the current
foreground process. The history can be optionally turned off, and when
job control is in play, it will mistakenly record 'fg' as the current
process.

https://bugzilla.gnome.org/show_bug.cgi?id=711059