screen: Try harder to find a foreground process group member
For pipelined commands, it's possible that the process group leader,
whose PID is used as the group ID, has terminated but not the entire
group. eg., $ cat <file> | less.  Sadly, there's no nice user-space
API to list the members of a group, and crawling the entire /proc like
pgrep might be overkill if the system is overloaded.

Therefore, the next 20 PIDs after the group ID and the first 20 PIDs
starting from 2 are checked in the hope that one of them would belong
to the foreground process group.

Subsequent commits will use this to track the current foreground
process, instead of using the shell's history, to notify when a
long-running foreground process group terminates.

https://bugzilla.gnome.org/show_bug.cgi?id=711059